+++ app/src/main/java/de/tum/ecorp/reservationapp/resource/RestaurantResource.java
@@ -27,5 +27,5 @@ public interface RestaurantResource {
     List<Restaurant> getRestaurantsNearby(Location searchLocation, Double searchRadius);
     List<Restaurant> getRestaurantsFiltered(Map<Filter, Object> filters);
 
-    Restaurant getRestaurant(Long restaurantId);
+    Restaurant getRestaurant(Integer restaurantId);
 }
